Se agrega escena juego
