+++ assets/cfg/interface.json
@@ -14,5 +14,5 @@
         "g": 144,
         "b": 255
     },
-    "high_score_max_value": 1000
+    "high_score_max_value": 5000
 }
